+++ src/Commands/1_music/nightcore.js
@@ -3,7 +3,8 @@ const Command = require('../../Structures/Command');
 const queue = require('../../Data/queue');
 const downloader = require('../../Data/downloader');
 const { consoleLog } = require('../../Data/Log');
-const { emoji: { success, warning }, response: { wrongChannel, noMusic, invalidNumber } } = require('../../../config/config.json');
+const timeConventer = require('../../Data/time');
+const { emoji: { success, warning }, response: { wrongChannel, noMusic, invalidNumber }, player: { downloaderMaxTimeSeconds } } = require('../../../config/config.json');
 
 
 module.exports = new Command({
@@ -24,6 +25,7 @@ module.exports = new Command({
 			return message.channel.send(`${success} Nightcore disabled.`);
 		}
 
+		if (guildQueue.songs[0].seconds > downloaderMaxTimeSeconds) return message.channel.send(`${warning} Track is too long to download for nightcore. Maxiumum length is ${timeConventer(downloaderMaxTimeSeconds)}.`);
 		if (args[0] && isNaN(args[0])) return message.channel.send(`${warning} ${invalidNumber}`);
 		if (args[0] < 0.5 || args[0] > 2) return message.channel.send(`${warning} ${invalidNumber} (0.5 - 2)`);