+++ src/Commands/1_music/nowplaying.js
@@ -3,7 +3,7 @@ const Command = require('../../Structures/Command');
 const { EmbedBuilder, ReactionCollector } = require('discord.js');
 const timeConverter = require('../../Data/time');
 const queue = require('../../Data/queue');
-const { bot: { ownerID, prefix }, emoji: { success, warning, error }, response: { wrongChannel, noMusic }, player: { updateInterval } } = require('../../../config/config.json');
+const { bot: { ownerID, prefix }, emoji: { success, warning, error }, response: { wrongChannel, noMusic }, player: { updateIntervalMiliseconds } } = require('../../../config/config.json');
 const { homepage } = require('../../../package.json');
 
 
@@ -20,6 +20,8 @@ module.exports = new Command({
 
         if (!message.member.voice.channel || guildQueue.connection.joinConfig.channelId != message.member.voice.channel.id) return message.channel.send(`${warning} ${wrongChannel}`);
 
+        if (!guildQueue.songs.length) return message.channel.send(`${warning} The queue is empty.`);
+
         guildQueue.updater.emitter.emit('end');
 
         const { requester, requester2 } = guildQueue.songs[0];
@@ -27,24 +29,33 @@ module.exports = new Command({
         let embed = new EmbedBuilder()
             .setColor(0x3399FF)
             .setAuthor({
-                name: `${requester.tag}${requester2 && requester.id != requester2.id ? ` - ${requester2.tag}` : ''}`,
+                name: `${requester.username}${requester2 && requester.id != requester2.id ? ` - ${requester2.username}` : ''}`,
                 url: homepage,
                 iconURL: requester.displayAvatarURL({ size: 1024, dynamic: true })
             })
             .setTitle(guildQueue.songs[0].title)
             .setURL(guildQueue.songs[0].url)
             .setDescription(`:arrow_forward: :radio_button:${line.repeat(14)} \`[0:00/${guildQueue.songs[0].length}]\` :loud_sound:`)
-            .setFooter({ text: `Source: ${guildQueue.songs[0].source}`})
+            //.setFooter({ text: `Source: ${guildQueue.songs[0].source}`})
+            .addFields([ { name: ' ', value: `Source: ${guildQueue.songs[0].source}`, inline: false } ])
+
+        if (guildQueue.songs[1]) {
+            embed.addFields([ 
+                { name: ' ', value: ' ', inline: false },
+                { name: 'Next', value: `[**${guildQueue.songs[1].title}**](${guildQueue.songs[1].url})`, inline: false } ]);
+        }
 
         const response = await message.channel.send({ content: `${success} **Now Playing...**`, embeds: [embed] });
 
-        if (updateInterval <= 0) return;
+        if (updateIntervalMiliseconds <= 0) return;
 
+        let allEmoji;
         const react = async () => { 
             for (const emoji of emojiList) {
-                response.react(emoji); 
+                if ( (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.react(emoji);
                 await new Promise(resolve => setTimeout(resolve, 750));
-            } 
+            }
+            allEmoji = true;
         }
         react(); 
 
@@ -88,22 +99,29 @@ module.exports = new Command({
 
         });
 
-        guildQueue.updater.emitter.on('update', async (song, duration) => {
+        guildQueue.updater.emitter.on('update', async (song, duration, nextsong) => {
             if (args[0] && args[0].toLowerCase() == 'debug' && message.author.id == ownerID) console.log('npupdate', song, duration);
             track = song;
             const progress = Math.round( duration / song.seconds * 14)
             embed = new EmbedBuilder()
                 .setColor(0x3399FF)
                 .setAuthor({
-                    name: `${song.requester.tag}${song.requester2 && song.requester.id != song.requester2.id ? ` - ${song.requester2.tag}` : ''}`,
+                    name: `${song.requester.username}${song.requester2 && song.requester.id != song.requester2.id ? ` - ${song.requester2.username}` : ''}`,
                     url: homepage,
                     iconURL: song.requester.displayAvatarURL({ size: 1024, dynamic: true })
                 })
                 .setTitle(song.title)
                 .setURL(song.url)
                 .setDescription(`${guildQueue.player.state.status == 'playing' ? ':arrow_forward:' : ':pause_button:'} ${line.repeat(progress)}:radio_button:${ (14 - progress) >= 0 ? line.repeat(14 - progress) : 0} \`[${timeConverter(duration)}/${song.length}]\` :loud_sound:`)
-                .setFooter({ text: `Source: ${song.source}`})
+                //.setFooter({ text: `Source: ${song.source}`})
+                .addFields([ { name: ' ', value: `Source: ${song.source}`, inline: false } ])
             
+            if (nextsong) {
+                embed.addFields([ 
+                    { name: ' ', value: ' ', inline: false },
+                    { name: 'Next', value: `[**${nextsong.title}**](${nextsong.url})`, inline: false } ]);
+            }
+
             if ( (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.edit({ content: `${success} **Now Playing...**`, embeds: [embed] });
             else {
                 collector.stop();
@@ -120,7 +138,7 @@ module.exports = new Command({
                 .setDescription(`:stop_button: ${line.repeat(15)} \`[terminated]\` :loud_sound:`)
                 .setFooter(null)
 
-            if ( (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.edit({ content: `${success} **Now Playing...**`, embeds: [embed] });
+            if ( (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.edit({ content: `${success} **Player was terminated.**`, embeds: [embed] });
             
             collector.stop();
             guildQueue.updater.emitter.removeAllListeners('update');
@@ -129,6 +147,7 @@ module.exports = new Command({
 
         collector.on('end', async (_, reason) => {
 			if (reason.endsWith('Delete')) return;
+            while (!allEmoji) await new Promise(resolve => setTimeout(resolve, 100));
 			response.reactions.removeAll();
         });
 	}