+++ src/Commands/1_music/play.js
@@ -4,13 +4,23 @@ const { ReactionCollector } = require('discord.js');
 const ytdl = require('ytdl-core');
 const ytpl = require('ytpl');
 const ytsr = require('yt-search');
+const playdl = require('play-dl');
 const { readdirSync } = require('fs')
 
 const { consoleLog } = require('../../Data/Log');
 const timeConverter = require('../../Data/time');
 const queue = require('../../Data/queue');
-const { emoji: { success, info, warning, error, loading }, response: { missingArguments, noChannel, wrongChannel, afkChannel }, player: { playlistFolderPath, maxTime, forbiddenKeywords } } = require('../../../config/config.json');
+const { emoji: { success, info, warning, error, loading }, response: { missingArguments, noChannel, wrongChannel, afkChannel }, player: { playlistFolderPath, maxTimeSeconds, forbiddenKeywords, youtubeCookie } } = require('../../../config/config.json');
 
+
+playdl.getFreeClientID().then((clientID) => playdl.setToken({
+    soundcloud : {
+        client_id : clientID
+    },
+    youtube : {
+        cookie : youtubeCookie
+    }
+}));
 
 const emojiList = [ '📥', '🚫' ]
 
@@ -18,7 +28,7 @@ module.exports = new Command({
 	name: 'play',
 	aliases: [ 'p' ],
     syntax: 'play <YouTube URL | Playlist URL | Search Query | subcommand>',
-	description: 'Plays a song or a playlist. Use top (t) or now (n) to put the song at the top of the queue or to play it immediately.',
+	description: 'Plays a song or a playlist. Use top (t) or now (n) as firts argument to put the song at the top of the queue or to play it immediately.',
 	async run(message, args, client) {
         if (!message.member.voice.channel) return message.channel.send(`${warning} ${noChannel}`);
         if (message.member.voice.channel.id == message.guild.afkChannelId) return message.channel.send(`${warning} ${afkChannel}`);
@@ -39,14 +49,14 @@ module.exports = new Command({
         if ( [ 'top', 't', 'next', 'now', 'n' ].includes(args[0].toLowerCase()) ) position = args.shift();
 
         if (!guildQueue) {
-            queue.construct(message, songs);
+            queue.construct(message, songs.slice());   //slice() to copy the array instead of passing the reference
         }
 
         const finder = (keywords, requester2Id, ignoreAttList = false) => {
             return new Promise(async resolve => {
                 let requester2 = null;
                 if (requester2Id) requester2 = await client.users.fetch(requester2Id);
-        
+
                 if (ytpl.validateID(keywords[0]) && ytdl.validateURL(keywords[0]) && !ignoreAttList) {
                     const skipTitle = await loadVideo(keywords[0], requester2)
                     if (skipTitle === 1) {
@@ -55,25 +65,27 @@ module.exports = new Command({
                     }
                     resolve(true);
                     if ( (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.edit(`${success} Added **${songs[0].title}** (\`${songs[0].length}\`) to the queue\n${info} This track has a playlist attached. Select ${emojiList[0]} to load playlist.`);
-        
-        
+
+
+                    let allEmoji;
                     const react = async () => { 
                         for (const emoji of emojiList) {
+                            if ( (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.react(emoji);
                             await new Promise(resolve => setTimeout(resolve, 750));
-                            response.react(emoji); 
-                        } 
+                        }
+                        allEmoji = true;
                     }
                     react();
-                
+
                     const filter = (reaction, user) => emojiList.includes(reaction.emoji.name) && user.bot == false;
-        
+
                     const collector = new ReactionCollector(response, { filter, time: 15000 } );
-        
+
                     collector.on('collect', async reaction => {
                         if (reaction.count < 2) return;
-        
+
                         reaction.users.remove(message.author);
-                        
+
                         switch (reaction.emoji.name) {
                             case emojiList[0]:
                                 collector.stop();
@@ -82,18 +94,19 @@ module.exports = new Command({
                                     queue.push(message.guild.id, element);
                                 });
                                 break;
-        
+
                             case emojiList[1]:
                                 collector.stop();
                                 break;
                         }
                     });
-        
+
                     collector.on('end', async (_, reason) => {
                         if (reason.endsWith('Delete')) return;
+                        while (!allEmoji) await new Promise(resolve => setTimeout(resolve, 100));
                         response.reactions.removeAll();
                     });
-        
+
                 }
                 else if (ytpl.validateID(keywords[0])) {
                     await loadList(keywords[0], requester2);
@@ -106,10 +119,29 @@ module.exports = new Command({
                     }
                     resolve();
                 }
+                else if (await playdl.so_validate(keywords[0]) == 'playlist') {
+                    if ( (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.edit(`${error} Soundcloud playlists are not supported.`);
+                    return;
+                }
+                else if (await playdl.so_validate(keywords[0]) == 'track') {
+                    const result = await playdl.soundcloud(keywords[0]);
+                    if (result.durationInSec > maxTimeSeconds && maxTimeSeconds > 0) return message.channel.send(`${error} **${result.name}**[${timeConverter(result.durationInSec)}] is too long! Max length is **${timeConverter(maxTimeSeconds)}**.`);
+                    if (forbiddenKeywords.some(element => result.name.toLowerCase().normalize('NFD').replace(/[\u0300-\u036f]/g, "").includes(element))) return message.channel.send(`${error} **${result.name}** is forbidden!`);
+                    songs.push({
+                        title: result.name,
+                        url: result.url,
+                        length: timeConverter(result.durationInSec),
+                        seconds: result.durationInSec,
+                        source: result.user.name,
+                        requester: message.author,
+                        requester2: requester2
+                    });
+                    resolve();
+                }
                 else {
                     const result = await ytsr(keywords.join(' '));
                     if (result < 1) return message.channel.send(`${error} Error finding video.`);
-                    if (result.videos[0].seconds > maxTime && maxTime > 0) return message.channel.send(`${error} **${element.title}**[${timeConverter(result.videos[0].seconds)}] is too long! Max length is **${timeConverter(maxTime)}**.`);
+                    if (result.videos[0].seconds > maxTimeSeconds && maxTimeSeconds > 0) return message.channel.send(`${error} **${element.title}**[${timeConverter(result.videos[0].seconds)}] is too long! Max length is **${timeConverter(maxTimeSeconds)}**.`);
                     if (forbiddenKeywords.some(element => result.videos[0].title.toLowerCase().normalize('NFD').replace(/[\u0300-\u036f]/g, "").includes(element))) return message.channel.send(`${error} **${result.videos[0].title}** is forbidden!`);
                     songs.push({ 
                         title: result.videos[0].title,
@@ -124,13 +156,13 @@ module.exports = new Command({
                 }
             });
         }
-        
+
         const loadList = async (url, requester2, skipTitle) => {
             return new Promise(async resolve => {
                 const result = await ytpl(url, { hl: 'cs', gl: 'CZ', limit: Infinity });
                 for (const element of result.items) {
-                    if (element.durationSec > maxTime && maxTime > 0) {
-                        message.channel.send(`${error} **${element.title}**[${timeConverter(element.durationSec)}] is too long! Max length is **${timeConverter(maxTime)}**.`);
+                    if (element.durationSec > maxTimeSeconds && maxTimeSeconds > 0) {
+                        message.channel.send(`${error} **${element.title}**[${timeConverter(element.durationSec)}] is too long! Max length is **${timeConverter(maxTimeSeconds)}**.`);
                         await new Promise(resolve => setTimeout(resolve, 1250));  //wait to not get ratelimited by discord api (1250ms)
                         continue;
                     }
@@ -156,9 +188,9 @@ module.exports = new Command({
         }
         const loadVideo = async (url, requester2) => {
             return new Promise(async resolve => {
-                ytdl.getBasicInfo(url)
+                ytdl.getBasicInfo(url, { requestOptions: { headers: { cookie: youtubeCookie } } } )
                     .then(result => {
-                        if (result.videoDetails.lengthSeconds > maxTime && maxTime > 0) return message.channel.send(`${error} **${result.videoDetails.title}**[${timeConverter(result.videoDetails.lengthSeconds)}] is too long! Max length is **${timeConverter(maxTime)}**.`);
+                        if (result.videoDetails.lengthSeconds > maxTimeSeconds && maxTimeSeconds > 0) return message.channel.send(`${error} **${result.videoDetails.title}**[${timeConverter(result.videoDetails.lengthSeconds)}] is too long! Max length is **${timeConverter(maxTimeSeconds)}**.`);
                         if (forbiddenKeywords.some(element => result.videoDetails.title.toLowerCase().normalize('NFD').replace(/[\u0300-\u036f]/g, "").includes(element))) return message.channel.send(`${error} **${result.videoDetails.title}** is forbidden!`);
                         songs.push({ 
                             title: result.videoDetails.title,
@@ -179,7 +211,7 @@ module.exports = new Command({
                     });
             });
         }
-        
+
         playlistBreak: 
         if ( [ 'playlist', 'pl' ].includes(args[0].toLowerCase()) ) {
             const files = readdirSync(`${__dirname}/../../../config/${playlistFolderPath}`);
@@ -205,6 +237,9 @@ module.exports = new Command({
         }
 
         if (!guildQueue) {
+            songs.forEach(element => {
+                queue.push(message.guild.id, element);
+            });
             if (shuffle) queue.shuffle(message.guild.id, 0);
 
             queue.player(message.guild.id);
@@ -222,11 +257,12 @@ module.exports = new Command({
                 songs.forEach(element => {
                     queue.push(message.guild.id, element);
                 });
-                if (shuffle) queue.shuffle(message.guild.id, guildQueue.songs.length);
+                if (shuffle) queue.shuffle(message.guild.id, guildQueue.songs.length - songs.length);  //guildQueue.songs.length is reference so the length is including the newly pushed songs
             }
+
             if (songs.length == 1 && !dontEdit && (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.edit(`${success} Added **${songs[0].title}** (\`${songs[0].length}\`) ${[ 'now', 'n' ].includes(position) || guildQueue.songs.length == 1 ? `to begin playing` : `to the queue at position ${position ? '1' : guildQueue.songs.length - 1}` } `);
             else if (response.editable && !dontEdit && (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.edit(`${success} Added ${shuffle ? 'and shuffled ' : ''}**${songs.length}** tracks!`);
-            
+
             if (guildQueue.player.state.status == 'idle') {
                 await new Promise(resolve => setTimeout(resolve, 100));
                 if (guildQueue.player.state.status == 'idle') queue.player(message.guild.id);
@@ -236,4 +272,3 @@ module.exports = new Command({
         }
 	}
 });
-